+++ src/main/java/com/cburch/logisim/instance/InstanceTextField.java
@@ -45,6 +45,7 @@ public class InstanceTextField implements AttributeListener, TextFieldListener, 
   private int fieldY;
   private int halign;
   private int valign;
+  private boolean multiline;
 
   InstanceTextField(InstanceComponent comp) {
     this.comp = comp;
@@ -70,7 +71,7 @@ public class InstanceTextField implements AttributeListener, TextFieldListener, 
 
   private void createField(AttributeSet attrs, String text) {
     final var font = attrs.getValue(fontAttr);
-    field = new TextField(fieldX, fieldY, halign, valign, font);
+    field = new TextField(fieldX, fieldY, halign, valign, font, multiline);
     field.setText(text);
     field.addTextFieldListener(this);
   }
@@ -137,6 +138,17 @@ public class InstanceTextField implements AttributeListener, TextFieldListener, 
 
   void update(
       Attribute<String> labelAttr, Attribute<Font> fontAttr, int x, int y, int halign, int valign) {
+    update(labelAttr, fontAttr, x, y, halign, valign, false);
+  }
+
+  void update(
+      Attribute<String> labelAttr,
+      Attribute<Font> fontAttr,
+      int x,
+      int y,
+      int halign,
+      int valign,
+      boolean multiline) {
     final var wasReg = shouldRegister();
     this.labelAttr = labelAttr;
     this.fontAttr = fontAttr;
@@ -144,6 +156,11 @@ public class InstanceTextField implements AttributeListener, TextFieldListener, 
     this.fieldY = y;
     this.halign = halign;
     this.valign = valign;
+    if (field != null && this.multiline != multiline) {
+      field.removeTextFieldListener(this);
+      field = null;
+    }
+    this.multiline = multiline;
     final var shouldReg = shouldRegister();
     var attrs = comp.getAttributeSet();
     fontColor =
@@ -156,6 +173,10 @@ public class InstanceTextField implements AttributeListener, TextFieldListener, 
     if (wasReg && !shouldReg) attrs.removeAttributeListener(this);
 
     updateField(attrs);
+  }
+
+  boolean isMultiline() {
+    return multiline;
   }
 
   private void updateField(AttributeSet attrs) {