+++ CHANGES.md
@@ -15,6 +15,7 @@
   * Added Real-Time Clock component.
   * Added Floating Point Constant component.
   * Modified paste behavior to paste at current mouse location if it is on canvas.
+  * Added multiline Text Tool labels using Shift+Enter or multiline clipboard text.
   * Fixed a regression that caused TestVector to fail when the circuit had subcircuits.
   * Fixed TTL 7447 BI/RBO port to be an input/output port to allow cascading of blanking mode.
   * Added TTL 7476: dual J-K Flip-flop with preset and clear.